wip: move the singleton holder over to the core module
