--- malime.kitsu/src/main/java/com/chesire/malime/kitsu/SingletonHolder.kt
@@ -1,30 +1,0 @@
-package com.chesire.malime.kitsu
-
-// Reference: https://medium.com/@BladeCoder/kotlin-singletons-with-argument-194ef06edd9e
-/**
- * Class used to create Singletons with a passed in parameter while keeping thread safety.
- */
-open class SingletonHolder<out T, in A>(creator: (A) -> T) {
-    private var creator: ((A) -> T)? = creator
-    @Volatile
-    private var instance: T? = null
-
-    fun getInstance(arg: A): T {
-        val i = instance
-        if (i != null) {
-            return i
-        }
-
-        return synchronized(this) {
-            val i2 = instance
-            if (i2 != null) {
-                i2
-            } else {
-                val created = creator!!(arg)
-                instance = created
-                creator = null
-                created
-            }
-        }
-    }
-}
